Split build and dev tests
